+++ core/mixins.py
@@ -24,6 +24,10 @@ TEMPLATE_MAPPING = {
     'InternationalGuideLandingPage': 'core/guide_landing_page.html',
     'InternationalEUExitFormPage': 'euexit/international-contact-form.html',
     'InternationalEUExitFormSuccessPage': 'euexit/international-contact-form-success.html',  # noqa
+    'InternationalCapitalInvestLandingPage': 'core/capital_invest_landing_page.html',       # noqa
+    'CapitalInvestRegionPage': 'core/capital_invest_region_page.html',
+    'CapitalInvestRegionalSectorPage': 'core/capital_invest_sector_page.html',
+    'CapitalInvestOpportunityPage': 'core/capital_invest_opportunity_page.html'
 }
 
 FEATURE_FLAGGED_URLS_MAPPING = {
@@ -81,7 +85,6 @@ class CMSPageMixin:
         this avoids /international/article-slug showing an article page
         at a url where it shouldn't exist
         """
-
         if self.page['page_type'] != self.page_type:
             raise Http404
         return super().dispatch(request, *args, **kwargs)